requirements.txt
django
djangorestframework
django-rest-knox
psycopg2
vaderSentiment
tweepy
pandas
tqdm
xlrd
pyyaml
coreapi
scrapy
beautifulsoup4
newspaper3k
nltk